Nove project.py to crud directory
